refector role & schedule controller

+++ src/Modules/roles/role.service.ts
@@ -1,3 +1,4 @@
+import { Response } from "express";
 import { prisma } from "../../prismaClient";
 import ApiError from "../../utils/ApiError";
 import ApiFeatures from "../../utils/ApiFeatures";
@@ -6,11 +7,15 @@ export const roleExist = async (name: string) => {
   return prisma.role.findFirst({ where: { name } });
 };
 
-export const createRole = async (body: any) => {
-  return prisma.role.create({ data: body });
+export const createRole = async (res: Response, body: any) => {
+  if (await roleExist(body.name))
+    throw new ApiError("this role name already exist", 409);
+
+  const role = await prisma.role.create({ data: body });
+  return res.status(200).json(role);
 };
 
-export const listRole = async (query: any) => {
+export const listRole = async (res: Response, query: any) => {
   const apiFeatures = new ApiFeatures(prisma.role, query);
 
   await apiFeatures.filter().search("role"); // Specify the model name, 'user' in this case
@@ -20,10 +25,10 @@ export const listRole = async (query: any) => {
   // Execute the query and get the result and pagination
   const { result, pagination } = await apiFeatures.exec("role");
 
-  return {
-    result,
-    pagination,
-  };
+  return res.status(200).json({
+    data: result,
+    pagination: pagination,
+  });
 };
 
 export const getUser = async (userId: number) => {
@@ -37,10 +42,16 @@ export const getRole = async (roleId: string) => {
 };
 
 export const assignRoleToUser = async (
+  res: Response,
   userId: number,
   roleId: string,
   id: number
 ) => {
+  if (!(await getUser(userId))) {
+    throw new ApiError("user not found", 404);
+  } else if (!(await getRole(roleId))) {
+    throw new ApiError("role not found", 404);
+  }
   if (userId === id) {
     throw new ApiError("you not allow to change your Role ..!", 401);
   }
@@ -49,22 +60,24 @@ export const assignRoleToUser = async (
       userId,
     },
   });
-  return prisma.userRole.create({
+  await prisma.userRole.create({
     data: {
       userId,
       roleId: parseInt(roleId, 10),
     },
   });
+  return res.json({ message: "assigning role to user successfully" });
 };
 
-export const listRoleUser = async (id: number) => {
-  return prisma.userRole.findMany({
+export const listRoleUser = async (res: Response, id: number) => {
+  let all = await prisma.userRole.findMany({
     where: { userId: id },
     include: {
       user: true,
       role: true,
     },
   });
+  res.status(200).json(all);
 };
 
 export const getRoleById = async (id: number) => {
@@ -76,14 +89,23 @@ export const getRoleById = async (id: number) => {
   });
 };
 
-export const updateRole = async (id: number, body: any) => {
-  return prisma.role.update({
+export const updateRole = async (res: Response, id: number, body: any) => {
+  if (!(await getRoleById(id))) throw new ApiError("role not found");
+
+  if (await roleExist(body.name)) {
+    throw new ApiError("this role name already exist", 409);
+  }
+  await prisma.role.update({
     where: { id },
     data: body,
   });
+  return res.status(200).json({ message: "role updated successfully" });
 };
 
-export const DeleteRole = async (id: number) => {
+export const DeleteRole = async (res: Response, id: number) => {
+  if (!(await getRoleById(id))) {
+    throw new ApiError("role not found");
+  }
   await prisma.userRole.deleteMany({
     where: {
       roleId: id,
@@ -97,4 +119,5 @@ export const DeleteRole = async (id: number) => {
   await prisma.role.delete({
     where: { id },
   });
+  return res.status(200).json({ message: "role deleted successfully" });
 };
